left align
can we make it left aligned but only go as far left as the image is?

Sure! To align the `figcaption` to the left but ensure it doesn't extend beyond the width of the image, you can use the following CSS:

```css
article figure {
    display: flex;
    flex-direction: column;
    align-items: center; /* Center the image and caption */
}

article figure figcaption {
    font-size: 0.6em;
    line-height: 0.9;
    margin-top: 0.5em;
    text-align: left; /* Align text to the left */
    max-width: 100%; /* Ensure it doesn't exceed the image width */
}
```

This will center the image and the caption within the figure, while keeping the caption text left-aligned and constrained to the width of the image. Let me know if you need any further adjustments!