try more
Let's try a different approach to ensure the `figcaption` is only as wide as the `picture` element. We can use a wrapper around the `picture` and `figcaption` to control their width. Here's the updated CSS:

```css
article figure {
    display: flex;
    flex-direction: column;
    align-items: center;
}

article figure .media-wrapper {
    display: block;
    max-width: 100%;
    width: auto;
}

article figure picture {
    display: block;
    max-width: 100%;
    height: auto;
}

article figure figcaption {
    font-size: 0.6em;
    line-height: 0.9;
    margin-top: 0.5em;
    text-align: left;
    width: 100%;
    box-sizing: border-box;
    padding-left: 0;
}
```

And update your HTML to include the wrapper:

```html
<article id="article">
    <figure>
        <div class="media-wrapper">
            <picture>
                <img alt="a photograph of a street sign with an arrow going in both directions. that is below a sign for Palestine Ave." src="[redacted]">
            </picture>
            <figcaption>...</figcaption>
        </div>
        <!-- Other content -->
    </figure>
</article>
```

This should ensure that the `figcaption` is only as wide as the `picture` element. Let me know if this works for you!