update alamo
writing prompt: here is a controversial opinion - we only remember the alamo because we won the war. sacrifices, hardships, and suffering don't matter by themselves unless accompanied by or quickly followed through by success. write a long essay filled with facts and backing sources to make this point. Make it as long as possible. do not make this short. we will continue with other prompts on this. do not refer to yourself as a chat bot, open ai, llm, or anything like that. write fluidly as a human would but don't introduce unnecessary typos. be clear and precise. don't hallucinate. use facts. don't make up fake news. and most importantly, please put everything in your responses after this in three ticks like ```this``` and do NOT say things like "got it" or "feel free to bla bla bla" or "it looks like you are writing a letter. would you like some help?" no nonsense. no prompts. just answers. and remember do NOT hallucinate.