add prompt
fix spelling, grammer, add line breaks, and do general clean up where necessary. this is a transcript. also don't output anything other than what is in the text. don't ask how my day was, don't ask if there is anything you can do, don't say anything like "here is what you ordered" or any such nonsense. always put answers inside three ticks like ```this``` so it is easy to copy paste.

this video is brought to you by brilliant outside of the US in the rest of the Western World Donald Trump isn't the most popular politician his first term coincided with an unprecedented collapse in America's Global reputation at least amongst its allies and Yuga polling ahead of the 2024 election found that voters in the UK France Germany Italy Spain Sweden and Denmark would all vote for Harris by overwhelming margins if they were able to however one place where Trump really is popular is Israel Channel 12 polling from late October suggested that an overwhelming 66% majority of Israelis wanted Trump to win the presidency with just 177% of Israelis saying they preferred Harris and Israeli politicians reacted to his reelection with Jubilation in a post on X for instance Israeli Prime Minister Benjamin Netanyahu hailed the results as the greatest comeback ever while his far-right Finance Minister bezel smotrich announced that 202 would be the year of sovereignty in Judea and Samaria essentially implying that Israel plans to use Trump's return to Annex the West Bank which didn't go down well with the International Community so in this video we're going to take a look at why Trump is so popular in Israel what his Israel policy might end up looking like and why he might not end up being quite as staunchly pro-israel as Netanyahu and Co [Music] assume before we start if you haven't already please consider subscribing and ringing the bell to stay in the loop and be notified when we release new videos so let's get straight into it and start with the big question why are Israelis and the Israeli government so keen on Trump well it's in part because Netanyahu and the Israeli right think that Biden has been too critical of Israel and worry that Harris would be even more so but it's mainly because Trump was pretty pro-israel during his first term in office his Israel policies included mve the US Embassy to Jerusalem cutting funding to unrar the UN Agency for Palestinian refugees and reversing Decades of US policy by recognizing Israeli sovereignty over the Golden Heights which Israel seized from Syria in 1967 Trump's Administration even announced that the US no longer considered Israeli settlements in the West Bank to be in violation of international law going against the position implicitly endorsed by the Obama Administration in late 2016 and flirted with actually supporting Israel's formal annexation of the occupied parts of the West Bank in early 2020 in this context you can see why his reelection has triggered the response it has from Netanyahu and his allies but it's at least possible that Trump won't be quite as unequivocally pro- Israel as Netanyahu and Co apparently assume for starters Trump no longer has a great personal relationship with Netanyahu while they apparently got on pretty well during Trump's first term Trump really soured on Netanyahu Yahoo after he immediately bubbled up to Joe Biden once he won the 2020 election