add second prompt
fix spelling, grammer, add line breaks, and do general clean up where necessary. this is a transcript. also don't output anything other than what is in the text. don't ask how my day was, don't ask if there is anything you can do, don't say anything like "here is what you ordered" or any such nonsense. always put answers inside three ticks like ```this``` so it is easy to copy paste.

in an interview with Israeli journalist Barack Ravid in December 2021 Trump said BBY BBY could have stayed quiet he's made a terrible mistake there were also a couple of moments in the campaign where Trump sounded surprisingly ambiguous while he did spend most of the campaign insisting he did more for Israel than any other president and that he was Israel's best friend he also said that Israel was quote losing the pr War by releasing the most heinous most horrible tapes of buildings falling down Trump even appeared to agree with a genocide Joe chant at one of his rallies when he turned to the crowd and said they're not wrong but the big reason that Trump might clash with Netanyahu is that he's always branded himself as the dealmaker during the campaign he promised to end the wars and even said in April that Israel should quote get back to peace and stop killing people but Trump's enthusiasm for quick deals will put him at odds with with the Israeli leadership because well netanyahu's maximalist War AIMS in Gaza namely destroying Hamas whatever that means can't be achieved anytime soon Netanyahu is unlikely to agree to anything that falls short because a pause before that point would embarrass him and probably mean new elections which he'd lose there's a similar dynamic in Lebanon where Israel's war aims seem to be constantly expanding and it's hard to imagine the current Israeli government accepting anything short of the complete removal of Hezbollah basically if Trump pushes for a deal Netanyahu is going to resist because any pause in the fighting before Hamas and Hezbollah are somehow totally destroyed will probably lead to his political downfall interestingly a similar Dynamic nearly played out in Trump's first term when Trump tried to solve the Israel Palestine problem to do this he literally didn't talk to the Palestinians only the Israelis and he along with Netanyahu and his son-in-law Jared Kushner came up with an unsurprisingly pro-israel plan where Israel was basically allowed to keep the land in the West Bank that had been occupied by settlers and most of Jerusalem in practice this meant that Palestinians only got about 70% of the West Bank and the territory was completely non-contiguous this was widely perceived by both the Palestinians and the International Community as being unfair to Palestine and the Palestinian leadership did actually reject it but what's interesting is that the plan only fully collapsed because at least according to journalist Thomas Freedman the pro-settler right wiers in netanyahu's government also rejected it on the grounds that they worried that seeding any land at all to the Palestinians would pave the way to a Palestinian state which they saw as unacceptable in other words Trump tried to make a deal so asked Netanyahu what do you want Netanyahu said 30% of the West Bank keep all the settlements and basically all of Jerusalem and Trump said fine only for Netanyahu to say uh actually I can't accept my own deal n yahu got out of this diplomatic mess with the Abraham Accords which essentially ignored the Palestinian issue and instead focused on normalizing Israel's relations with other Arab countries but if doing a deal with Netanyahu on Palestine was difficult in Trump's first term it's going to be even more difficult in his second now to be clear we're not saying that Trump won't be pro- Israel ultimately no one knows what Trump's Israel policy is going to be and a lot will depend on how much influence his cabinet has given their over whelmingly pro-israel especially his secretary of state Marco Rubio and his Israel Ambassador Mike Hooke who to give you a sense of quite how Pro Israeli is once said there's really no such thing as a Palestinian and there's no such thing as the West Bank it's Judea and Samaria there is no such thing as settlements their communities their neighborhoods their cities there's no such thing as an occupation nonetheless the point we're trying to make is that if Trump wants to protect his image as the non-nonsense dealmaker who keeps the world World in check he might well push for a deal in the Middle East and this probably will put him at odds with both Netanyahu and much of the Israeli right now as a tldr viewer it's clear that you have a thirst for learning after all why else would you have watched a 10-minute video explaining Trump's Israel policy now if you want to continue your Learning Journey you really should check out brilliant brilliant is the online learning platform where you learn by doing with thousands of interactive lessons in maths data analysis programming and AI these Hands-On lessons have you comparing circuits to understand voltage and current playing snooker to learn the rules of collision planning your itinery for an Intergalactic Music Festival on A Spacetime diagram and more that way you'll build your natural intuition while gaining deep knowledge of scientific principles in fact brilliant actually recently launched tons of new data science content all of which uses real world data to help you see Trends and make better informed decisions and don't worry you don't need to be a world-renowned scientist to understand these courses either Brilliance courses are designed for Learners of any level found out more and to try everything brilliant has to offer for free for a full 30 days visit brilliant.org tldr or click the link in the description you'll also get 20% off an annual premium subscription thanks for your support and thanks to brilliant for supporting this video