add metadata
Here are some suggestions for the title and description of your HTML document:

### Title:
```
The Fabulous Origin of Flair: A Tribute to Ric Flair
```

### Description:
```
Discover the outrageous and hilarious tale of how Ric Flair became the origin of the word "Flair." From dazzling robes to unforgettable antics, this whimsical story will have you laughing and appreciating the true meaning of style and originality. Join us in celebrating the legend himself!
```

Feel free to modify these suggestions to better fit your vision!